T651 sync CRM

Approved-by: Jack Tanner <jack@jackandtheblockstalk.com>

+++ server/test/services/zoho/zoho.spec.js
@@ -1,10 +1,9 @@
+const server = require('../../../server');
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const mongo = require('../../helpers/mongo');
 
-
-const server = require('../../../server');
-const users = require('../../../model/mongoose/users');
+const settings = require('../../../settings');
 const zoho = require('../../../controller/services/zoho/zoho');
 
 const assert = chai.assert;
@@ -16,29 +15,35 @@ describe('zoho tokens', function () {
     this.timeout(7000);
 
     afterEach(async function () {
-        console.log('dropping database');
-        await mongo.drop();
+        if (settings.ENVIRONMENT === 'test-all') {
+            console.log('dropping database');
+            await mongo.drop();
+        }
     })
 
     describe('using Zoho CRM API', function () {
 
         it('should get a new auth token and call the API', async function () {
-            await zoho.initialize();
-            await zoho.generateAuthTokenfromRefreshToken();
+            if (settings.ENVIRONMENT === 'test-all') {
+                await zoho.initialize();
+                await zoho.generateAuthTokenfromRefreshToken();
 
-            const contacts = await zoho.contacts.getMany({
-                params: {
-                    page: 0,
-                    per_page: 1
-                }
-            });
+                const contacts = await zoho.contacts.getMany({
+                    params: {
+                        page: 0,
+                        per_page: 1
+                    }
+                });
 
-            assert(contacts.length === 1, "No contacts founds");
+                assert(contacts.data.length === 1, "No contacts founds");
+            }
         })
 
         it('should get new auth token', async function () {
-            await zoho.initialize();
-            // await zoho.generateAuthTokens();
+            if (settings.ENVIRONMENT === 'test-all') {
+                await zoho.initialize();
+                // await zoho.generateAuthTokens();
+            }
         })
     })
 });
